# Better ARM_NEON implementation for R4 quants